requirements.txt
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
plotly==5.15.0
scikit-learn==1.3.0
tensorflow==2.13.0
ccxt==3.1.54
pytest==7.4.0
pytest-cov==4.1.0
flask==2.3.3
sqlalchemy==2.0.20
redis==4.6.0
pyyaml==6.0.1
openpyxl==3.1.2
xlrd==2.0.1
h5py==3.9.0
python-dotenv==1.0.0
yfinance==0.2.28 